requirements.txt
requests
redis>=4.2.0
python-dotenv
redisvl
